+++ client/src/App.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import Home from './components/Home'
 import LeetCodeWrapped from './components/LeetCodeWrapped'
-import Worth from './components/Worth'
+import Worth from './components/Worth/Worth'
 
 const App = () => {
   return (
